fix(automate): fix reviews comments
